messages now deleted on hang up - may make this an option message name cleared always after return to main prompt to avoid accidental delete

+++ trunk/asterisk-common/agi-bin/Lifeline.pm
@@ -386,8 +386,10 @@ sub clean_up_msgs {
 		if ($_->{deleted}) {
 			my $mname = $_->{msg}.".".$ll->{rectype};
 			my $dname = $_->{msg}.".deleted.".$ll->{rectype};
-			unlink $dname;
-			move $mname, $dname or print STDERR $!;
+			unlink $dname if -f $dname;
+			if (-f $mname) {
+				move $mname, $dname or print STDERR "$mname -> $dname: $!";
+			}
 		}
 	}
 }
